Scanner sc ADDED

+++ src/area.java
@@ -9,7 +9,7 @@ public class area {
 
 //volume of sphere
         Scanner sc = new Scanner(System.in);
-        double  r = sc.nextInt();
+        double  r = sc.nextDouble();
         double areOfSphere = 4.0 / 3.0 * Math.PI * r * r * r;
 //        double areOfSphere2 = 4.0 / 3.0 * Math.PI * Math.pow(r ,3);
         System.out.println(areOfSphere);
